Move ImageInfo to common package
