+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/MetadataExt.kt
@@ -1,17 +1,19 @@
 package com.mapbox.search.base
 
+import com.mapbox.search.base.core.CoreImageInfo
 import com.mapbox.search.base.core.CoreOpenHours
 import com.mapbox.search.base.core.CoreOpenMode
 import com.mapbox.search.base.core.CoreOpenPeriod
 import com.mapbox.search.base.core.CoreParkingData
 import com.mapbox.search.base.utils.printableName
+import com.mapbox.search.common.metadata.ImageInfo
 import com.mapbox.search.common.metadata.OpenHours
 import com.mapbox.search.common.metadata.OpenPeriod
 import com.mapbox.search.common.metadata.ParkingData
 import com.mapbox.search.common.metadata.WeekDay
 import com.mapbox.search.common.metadata.WeekTimestamp
 
-public fun CoreOpenHours.mapToPlatform(): OpenHours? = when (mode) {
+fun CoreOpenHours.mapToPlatform(): OpenHours? = when (mode) {
     CoreOpenMode.ALWAYS_OPEN -> OpenHours.AlwaysOpen
     CoreOpenMode.TEMPORARILY_CLOSED -> OpenHours.TemporaryClosed
     CoreOpenMode.PERMANENTLY_CLOSED -> OpenHours.PermanentlyClosed
@@ -28,7 +30,7 @@ public fun CoreOpenHours.mapToPlatform(): OpenHours? = when (mode) {
     }
 }
 
-public fun OpenHours.mapToCore() = when (this) {
+fun OpenHours.mapToCore() = when (this) {
     OpenHours.AlwaysOpen -> CoreOpenHours(CoreOpenMode.ALWAYS_OPEN, emptyList())
     OpenHours.TemporaryClosed -> CoreOpenHours(CoreOpenMode.TEMPORARILY_CLOSED, emptyList())
     OpenHours.PermanentlyClosed -> CoreOpenHours(CoreOpenMode.PERMANENTLY_CLOSED, emptyList())
@@ -36,7 +38,7 @@ public fun OpenHours.mapToCore() = when (this) {
     else -> error("Unknown OpenHours subclass: ${javaClass.printableName}.")
 }
 
-public fun CoreOpenPeriod.mapToPlatform() = OpenPeriod(
+fun CoreOpenPeriod.mapToPlatform() = OpenPeriod(
     open = WeekTimestamp(
         day = weekDayFromCore(openD),
         hour = openH.toInt(),
@@ -49,22 +51,34 @@ public fun CoreOpenPeriod.mapToPlatform() = OpenPeriod(
     ),
 )
 
-public fun OpenPeriod.mapToCore() = CoreOpenPeriod(
+fun OpenPeriod.mapToCore() = CoreOpenPeriod(
     open.day.internalRawCode, open.hour.toByte(), open.minute.toByte(),
     closed.day.internalRawCode, closed.hour.toByte(), closed.minute.toByte(),
 )
 
-public fun CoreParkingData.mapToPlatform() = ParkingData(
+fun CoreParkingData.mapToPlatform() = ParkingData(
     totalCapacity = capacity,
     reservedForDisabilities = forDisabilities
 )
 
-public fun ParkingData.mapToCore() = CoreParkingData(
+fun ParkingData.mapToCore() = CoreParkingData(
     totalCapacity,
     reservedForDisabilities
 )
 
-public fun weekDayFromCore(dayCode: Byte): WeekDay {
+fun weekDayFromCore(dayCode: Byte): WeekDay {
     return WeekDay.values().firstOrNull { it.internalRawCode == dayCode }
         ?: throw IllegalArgumentException("Unknown day code (=$dayCode) from Core SDK.")
 }
+
+fun CoreImageInfo.mapToPlatform(): ImageInfo = ImageInfo(
+    url = url,
+    width = width,
+    height = height
+)
+
+fun ImageInfo.mapToCore(): CoreImageInfo = CoreImageInfo(
+    url,
+    width,
+    height
+)
