+++ MapboxSearch/sdk-common/api/sdk-common.api
@@ -420,6 +420,19 @@ public final class com/mapbox/search/common/concurrent/SearchSdkMainThreadWorker
 	public final fun setDelegate (Lcom/mapbox/search/common/concurrent/MainThreadWorker;)V
 }
 
+public final class com/mapbox/search/common/metadata/ImageInfo : android/os/Parcelable {
+	public static final field CREATOR Landroid/os/Parcelable$Creator;
+	public fun <init> (Ljava/lang/String;II)V
+	public fun describeContents ()I
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getHeight ()I
+	public final fun getUrl ()Ljava/lang/String;
+	public final fun getWidth ()I
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+	public fun writeToParcel (Landroid/os/Parcel;I)V
+}
+
 public abstract class com/mapbox/search/common/metadata/OpenHours : android/os/Parcelable {
 }
