+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/SearchResultMetadata.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import com.mapbox.search.base.core.CoreResultMetadata
 import com.mapbox.search.base.mapToCore
 import com.mapbox.search.base.mapToPlatform
+import com.mapbox.search.common.metadata.ImageInfo
 import com.mapbox.search.common.metadata.OpenHours
 import com.mapbox.search.common.metadata.ParkingData
 import kotlinx.parcelize.IgnoredOnParcel
