+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/SearchResultMetadataTest.kt
@@ -2,6 +2,7 @@ package com.mapbox.search
 
 import com.mapbox.search.base.core.CoreResultMetadata
 import com.mapbox.search.base.mapToCore
+import com.mapbox.search.common.metadata.ImageInfo
 import com.mapbox.search.common.metadata.OpenHours
 import com.mapbox.search.common.metadata.OpenPeriod
 import com.mapbox.search.common.metadata.ParkingData
